requirements.txt
spacy-nightly>=2.1.0a5
stanfordnlp>=0.1.0,<0.2.0
